Support for while-pattern-guard; Bug fix ESPER-468  IllegalArgumentException while parsing "B until not B" pattern ; bug fix ESPER-464  Add an engine-level configuration parameter for default AccessorStyle
git-svn-id: http://svn.codehaus.org/esper/esper/branches/enhancements350@2105 d91eb1b3-1215-0410-9718-d41490442091

+++ esper/src/main/java/com/espertech/esper/pattern/EvalFilterStateNode.java
@@ -184,7 +184,9 @@ public final class EvalFilterStateNode extends EvalStateNode implements FilterHa
 
     private void stopFiltering()
     {
-        context.getFilterService().remove(handle);
+        if (handle != null) {
+            context.getFilterService().remove(handle);
+        }
         handle = null;
         isStarted = false;
         long filtersVersion = context.getFilterService().getFiltersVersion();
